vim mouse events in insert mode improvement
Turns out I don't like the mouse moving the cursor while in insert mode.
Probably because I lost mental context of what I was doing if I was
in a different window and I'm clicking back to vim.

So now switch to normal mode and set a jump point. That way I can just
either:
  * Switch back to insert mode where I am.
  * <C-o> to return to where the insert point previously was.