Add support for more paths
Decided I wanted to allow paths with two separators, even if no linenum.
Ending with a linenum is always a match. So now matches are:
* file:23
* ~/path/filename
* /path/filename
* path/path/filename

The more obvious ones still match, of course:

* path/file:23:45
* "~/path with spaces/file":23:45

We intentionally do not match single separators

* (No match) path/filename

If we allowed that, the regex could be shorter, so there's that.